\name{CMRCT-package}
\alias{CMRCT-package}
\alias{CMRCT}
\docType{package}

\title{
Capture-Mark-Recapture in Continuous Time
}

\description{
A suite of R routines to fit Capture-Mark-Recapture models in
continuous time. 
}

\details{
\tabular{ll}{
Package: \tab CMRCT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-29\cr
License: \tab GPL-2\cr
}

}

\author{Hugues Santin-Janin, David Fouchet

  Maintainer: Hugues Santin-Janin <hugues_sj@yahoo.fr>
}

\references{ Fouchet D., Santin-Janin H., Sauvage F., Yoccoz N.-G. &
Pontier D. (submitted) An R package for analyzing survival using
continuous-time open capture-recapture models: an example with feral cats on sub-Antarctic islands.}

\keyword{package}
