tototo
\name{CMRCT-package}
\alias{CMRCT-package}
\alias{CMRCT}
\docType{package}

\title{
Capture-Mark-Recapture in Continuous Time
}

\description{
A suite of R routines to fit Capture-Mark-Recapture models in
continuous time. 
}

\details{
\tabular{ll}{
Package: \tab CMRCT\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-09-29\cr
License: \tab GPL-2\cr
}

}

\author{H. Santin-Janin, D. Fouchet

  Maintainer: H Santin-Janin <hugues.santin-janin@oncfs.gouv.fr>
}

\references{ Fouchet D., Santin-Janin H., Sauvage F., Yoccoz N.-G. &
Pontier D. (submitted) An R package for analysing survival from
capture-mark-recapture non-standard protocols using continuous time
capture-recapture models.}

\keyword{package}
